Created MSOA population employment checks class
Created class (PopEmpComparator) for reading inputs  for the MSOA
population/employment data and comparing it with the output data.
The class has functionality for perfoming comparison at the column total
level with placeholder methods for the other comparisons.
Also added test function for testing the comparison class on outputs from
a previous run of EFS.