Added functionality to write comparisons to excel
Updated write_comparisons to write comparison summaries to spreadsheet
instead of CSVs (default) with parameter to write to CSVs instead.
Added function which is used by write_comparisons to format columns in
excel.
Added function to utils which checks if a file can be written to, used
by write_comparisons to avoid PermissionErrors.
Installed openpyxl into python environment and updated the Pipfiles.