Added Car RUC cost change
Added functionality for Car road user cost change calculations which
apply the cost change directly to the generlised cost so setup the
calculation to manage applying any change to GC directly may be useful
for future cost changes.
Updated test elasticities to match those in the evidence spreadsheet.
Updated tests to use new answers from spreadsheet with a tolerance of
10^-4 to account for rounding inconsistencies in Excel vs Python.