Removed hardcoded GC parameters, now reads from file
Added argument to `ElasticityModel` for path to the gc parameters CSV
file, which should contain parameters for all years and modes.
Changed `ElasticityModel._check_folders` method to `_check_paths` so
that it  can check paths to files as well as folders, used for checking
new parameter.
Reads GC parameters in the `apply_all` method and passes it to
`apply_elasticities`, instead of being hardcoded.
Updated `TestElasticityModel` to create gc parameters test file and pass
it to `apply_elasticities` method.
Changed keys for the gc parameters to 'vt' and 'vc', from 'vot'/'voc' to
match the keys that are used in `ElasticityModel`, updated
`TestGCParameters` to reflect this change.